+++ src/Settings/PreferencesWindow.cpp
@@ -101,9 +101,9 @@ PreferencesWindow::PreferencesWindow(QWidget *parent) : QMainWindow(parent)
 
     addPage("Language/General", {"Default Language"});
 
-    addPage("Language/Commands",
-            {"C++/Compile Command", "C++/Run Arguments", "Java/Compile Command", "Java/Run Arguments",
-             "Java/Run Command", "Python/Run Arguments", "Python/Run Command"});
+    addPage("Language/Commands", {"C++/Compile Command", "C++/Executable File Path", "C++/Run Arguments",
+                                  "Java/Compile Command", "Java/Run Arguments", "Java/Run Command", "Java/Class Name",
+                                  "Python/Run Arguments", "Python/Run Command"});
 
     addPage("Language/Code Template", {"C++/Template Path", "Java/Template Path", "Python/Template Path"});
 
@@ -120,7 +120,8 @@ PreferencesWindow::PreferencesWindow(QWidget *parent) : QMainWindow(parent)
 
     addPage("Actions/General", {"Hot Exit/Enable"});
 
-    addPage("Actions/Save", {"Auto Save", "Save Faster", "Auto Format", "Save File On Compilation", "Save File On Execution", "Save Tests"});
+    addPage("Actions/Save", {"Auto Save", "Save Faster", "Auto Format", "Save File On Compilation",
+                             "Save File On Execution", "Save Tests"});
 
     addPage("Extensions/Clang Format", {"Clang Format/Path", "Clang Format/Style"}, false);