+++ doc/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Now it's optional to save the file on compilation and execution or not.
+- Now you can choose the path of the executable file of C++. (#271)
 
 ### Fixed